fix stuff, probably next do tests
