--- src/yo/01-basic-table/04-Combination.stories.js
@@ -1,62 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { categoryName } from './_base';
-
-import moment from 'moment';
-// In your setup, replace "../../" with "react-bs-datatable".
-import Datatable from '../../';
-
-const header = [
-  {
-    title: 'Username (filterable)',
-    prop: 'username',
-    sortable: true,
-    filterable: true
-  },
-  { title: 'Name', prop: 'realname', sortable: true },
-  { title: 'Location', prop: 'location' },
-  { title: 'Last Updated', prop: 'date', sortable: true }
-];
-
-const body = Array.from(new Array(57), () => {
-  const rd = (Math.random() * 10).toFixed(1);
-
-  if (rd > 0.5) {
-    return {
-      username: 'i-am-billy',
-      realname: `Billy ${rd}`,
-      location: 'Mars',
-      date: moment()
-        .subtract(1, 'days')
-        .format('Do MMMM YYYY')
-    };
-  }
-
-  return {
-    username: 'john-nhoj',
-    realname: `John ${rd}`,
-    location: 'Saturn',
-    date: moment()
-      .subtract(2, 'days')
-      .format('Do MMMM YYYY')
-  };
-});
-
-const onSortFunction = {
-  date(columnValue) {
-    // Convert the string date format to UTC timestamp
-    // So the table could sort it by number instead of by string
-    return moment(columnValue, 'Do MMMM YYYY').valueOf();
-  }
-};
-
-storiesOf(categoryName, module).add('Filter, sort, and pagination', () => (
-  <Datatable
-    tableHeaders={header}
-    tableBody={body}
-    rowsPerPage={5}
-    rowsPerPageOption={[5, 10, 15, 20]}
-    initialSort={{ prop: 'username', isAscending: true }}
-    onSort={onSortFunction}
-  />
-));
