--- src/yo/01-basic-table/08-styling.css
@@ -1,3 +1,0 @@
-.custom-table > .tbody > .tbody-tr:hover {
-  background: #fafafa;
-}
