--- src/yo/01-basic-table/_base.js
@@ -1,1 +1,0 @@
-export const categoryName = 'Getting Started';
