--- src/yo/02-advanced/00-Extending.stories.js
@@ -1,199 +1,0 @@
-/**
- * This is a story for an extended table.
- * If you want to personally customize how the table looks, like the positioning of filters/pagination, you can do it here.
- * You will need to import the components and apply them in the render function.
- * If you are about to extend your table, do not forget to wrap it with memo.
- */
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { categoryName } from './_base';
-
-import moment from 'moment';
-
-// In your setup, replace "../../" with "react-bs-datatable"
-import {
-  Pagination,
-  PaginationOpts,
-  TableHeader,
-  TableBody,
-  Filter,
-  useDatatableLifecycle,
-  shouldTableUpdate
-} from '../../';
-
-const CustomTable = React.memo(props => {
-  const {
-    data,
-    rowsPerPageOption,
-    tableHeaders,
-    onChangeFilter,
-    onPageNavigate,
-    classes,
-    onRowsPerPageChange,
-    onSortChange,
-    tableClass,
-    labels,
-    filterable,
-    filterText,
-    rowsPerPage,
-    currentPage,
-    sortedProp,
-    maxPage,
-    Components
-  } = useDatatableLifecycle(props);
-
-  return (
-    <>
-      <Components.Row className="controlRow__root">
-        <Components.Col xs="12">
-          <Filter
-            classes={classes}
-            tableHeaders={tableHeaders}
-            placeholder={labels.filterPlaceholder}
-            onChangeFilter={onChangeFilter}
-            filterText={filterText}
-            filterable={filterable}
-            components={{
-              Adornment: Components.Adornment,
-              Button: Components.Button,
-              ClearIcon: Components.ClearIcon,
-              FormControl: Components.FormControl,
-              InputGroup: Components.InputGroup
-            }}
-          />
-        </Components.Col>
-      </Components.Row>
-      <Components.Row>
-        <Components.Col xs="12">
-          <Components.Table className={tableClass}>
-            <TableHeader
-              classes={classes}
-              tableHeaders={tableHeaders}
-              sortedProp={sortedProp}
-              onSortChange={onSortChange}
-              components={{
-                TableHead: Components.TableHead,
-                TableCell: Components.TableCell,
-                TableRow: Components.TableRow
-              }}
-            />
-            <TableBody
-              classes={classes}
-              tableHeaders={tableHeaders}
-              labels={labels}
-              data={data}
-              components={{
-                TableBody: Components.TableBody,
-                TableCell: Components.TableCell,
-                TableRow: Components.TableRow
-              }}
-            />
-          </Components.Table>
-        </Components.Col>
-      </Components.Row>
-      <Components.Row className="controlRow__root bottom">
-        <Components.Col xs={12} md={4} />
-        <Components.Col xs={12} md={4}>
-          <PaginationOpts
-            classes={classes}
-            labels={labels}
-            onRowsPerPageChange={onRowsPerPageChange}
-            rowsPerPage={rowsPerPage}
-            rowsPerPageOption={rowsPerPageOption}
-            components={{
-              Form: Components.Form,
-              FormGroup: Components.FormGroup,
-              FormControl: Components.FormControl
-            }}
-          />
-        </Components.Col>
-        <Components.Col xs={12} md={4} className="text-right">
-          <Pagination
-            classes={classes}
-            data={data}
-            rowsPerPage={rowsPerPage}
-            currentPage={currentPage}
-            onPageNavigate={onPageNavigate}
-            labels={labels}
-            maxPage={maxPage}
-            components={{
-              Button: Components.Button,
-              ButtonGroup: Components.ButtonGroup
-            }}
-          />
-        </Components.Col>
-      </Components.Row>
-    </>
-  );
-}, shouldTableUpdate);
-
-const header = [
-  {
-    title: 'Username (filterable)',
-    prop: 'username',
-    sortable: true,
-    filterable: true
-  },
-  { title: 'Name', prop: 'realname', sortable: true },
-  {
-    title: 'Name Uppercased',
-    prop: 'realnameuppercase',
-    cell: row => row.realname.toUpperCase()
-  },
-  { title: 'Location', prop: 'location' },
-  { title: 'Last Updated', prop: 'date', sortable: true }
-];
-
-const body = Array.from(new Array(57), () => {
-  const rd = (Math.random() * 10).toFixed(1);
-
-  if (rd > 0.5) {
-    return {
-      username: 'i-am-billy',
-      realname: `Billy ${rd}`,
-      location: 'Mars',
-      date: moment()
-        .subtract(1, 'days')
-        .format('Do MMMM YYYY')
-    };
-  }
-
-  return {
-    username: 'john-nhoj',
-    realname: `John ${rd}`,
-    location: 'Saturn',
-    date: moment()
-      .subtract(2, 'days')
-      .format('Do MMMM YYYY')
-  };
-});
-
-const onSortFunction = {
-  date(columnValue) {
-    // Convert the string date format to UTC timestamp
-    // So the table could sort it by number instead of by string
-    return moment(columnValue, 'Do MMMM YYYY').valueOf();
-  }
-};
-
-const customLabels = {
-  first: '<<',
-  last: '>>',
-  prev: '<',
-  next: '>',
-  show: 'Display',
-  entries: 'rows',
-  noResults: 'There are no data to be displayed'
-};
-
-storiesOf(categoryName, module).add('Extending the Table', () => (
-  <CustomTable
-    tableHeaders={header}
-    tableBody={body}
-    rowsPerPage={5}
-    rowsPerPageOption={[5, 10, 15, 20]}
-    initialSort={{ prop: 'username', isAscending: true }}
-    onSort={onSortFunction}
-    labels={customLabels}
-  />
-));
