--- src/yo/02-advanced/_base.js
@@ -1,1 +1,0 @@
-export const categoryName = 'Advanced Guides';
